+++ src/account-abstraction/auth.ts
@@ -1,3 +1,5 @@
+import { AxiosRequestConfig } from 'axios';
+
 import { XellarEWBase } from '../base';
 import { BaseHttpResponse } from '../types/http';
 import { handleError, XellarError } from '../utils/error';
@@ -9,16 +11,23 @@ export class XellarAAAuth extends XellarEWBase {
    * @param options - Options for creating a smart account
    * @returns The created smart account
    */
-  async createAccount({ owner }: AACreateAccountOptions) {
+  async createAccount(
+    { owner }: AACreateAccountOptions,
+    config?: AxiosRequestConfig,
+  ) {
     try {
       const response = await this.aaInstance.post<
         BaseHttpResponse<AACreateAccountResponse>
-      >('/smart-account', {
-        owner: {
-          id: owner.id,
-          provider: owner.provider,
+      >(
+        '/smart-account',
+        {
+          owner: {
+            id: owner.id,
+            provider: owner.provider,
+          },
         },
-      });
+        config,
+      );
 
       return response.data.data;
     } catch (error) {