+++ src/account-abstraction/user-op/create/send-native.ts
@@ -1,3 +1,5 @@
+import { AxiosRequestConfig } from 'axios';
+
 import { XellarEWBase } from '../../../base';
 import { BaseHttpResponse } from '../../../types/http';
 import { handleError, XellarError } from '../../../utils/error';
@@ -18,7 +20,7 @@ export class XellarAACreateSendNative extends XellarEWBase {
    * @example
    *
    * ```typescript
-   * const result = await sdk.aa.create.sendNative({
+   * const result = await sdk.accountAbstraction.create.sendNative({
    *   accountId: "67959f7927373a6808679de2",
    *   to: "0x...",
    *   amount: "1000000000000000000"
@@ -35,15 +37,20 @@ export class XellarAACreateSendNative extends XellarEWBase {
    */
   async sendNative(
     options: EstimateSendNativeOptions,
+    config?: AxiosRequestConfig,
   ): Promise<CreateUserOpResponse> {
     try {
       const response = await this.aaInstance.post<
         BaseHttpResponse<CreateUserOpResponse>
-      >('/userOp/sendCoin/hash', {
-        accountId: options.accountId,
-        to: options.to,
-        amount: options.amount,
-      });
+      >(
+        '/userOp/sendCoin/hash',
+        {
+          accountId: options.accountId,
+          to: options.to,
+          amount: options.amount,
+        },
+        config,
+      );
 
       return response.data.data;
     } catch (error) {