+++ src/account-abstraction/user-op/create/send-token.ts
@@ -1,3 +1,5 @@
+import { AxiosRequestConfig } from 'axios';
+
 import { XellarEWBase } from '../../../base';
 import { BaseHttpResponse } from '../../../types/http';
 import { handleError, XellarError } from '../../../utils/error';
@@ -19,7 +21,7 @@ export class XellarAACreateSendToken extends XellarEWBase {
    * @example
    *
    * ```typescript
-   * const result = await sdk.aa.create.sendToken({
+   * const result = await sdk.accountAbstraction.create.sendToken({
    *   accountId: "67959f7927373a6808679de2",
    *   tokenAddress: "0x...",
    *   to: "0x...",
@@ -37,16 +39,21 @@ export class XellarAACreateSendToken extends XellarEWBase {
    */
   async sendToken(
     options: EstimateSendTokenOptions,
+    config?: AxiosRequestConfig,
   ): Promise<CreateUserOpResponse> {
     try {
       const response = await this.aaInstance.post<
         BaseHttpResponse<CreateUserOpResponse>
-      >('/userOp/sendToken/hash', {
-        accountId: options.accountId,
-        tokenAddress: options.tokenAddress,
-        to: options.to,
-        amount: options.amount,
-      });
+      >(
+        '/userOp/sendToken/hash',
+        {
+          accountId: options.accountId,
+          tokenAddress: options.tokenAddress,
+          to: options.to,
+          amount: options.amount,
+        },
+        config,
+      );
 
       return response.data.data;
     } catch (error) {