+++ src/account-abstraction/user-op/create/sign-transaction.ts
@@ -1,3 +1,5 @@
+import { AxiosRequestConfig } from 'axios';
+
 import { XellarEWBase } from '../../../base';
 import { BaseHttpResponse } from '../../../types/http';
 import { handleError, XellarError } from '../../../utils/error';
@@ -19,7 +21,7 @@ export class XellarAACreateSignTransaction extends XellarEWBase {
    * @example
    *
    * ```typescript
-   * const result = await sdk.aa.create.signTransaction({
+   * const result = await sdk.accountAbstraction.create.signTransaction({
    *   accountId: "67959f7927373a6808679de2",
    *   to: "0x...",
    *   value: "0",
@@ -37,16 +39,21 @@ export class XellarAACreateSignTransaction extends XellarEWBase {
    */
   async signTransaction(
     options: EstimateSignTransactionOptions,
+    config?: AxiosRequestConfig,
   ): Promise<CreateUserOpResponse> {
     try {
       const response = await this.aaInstance.post<
         BaseHttpResponse<CreateUserOpResponse>
-      >('/userOp/signTransaction/hash', {
-        accountId: options.accountId,
-        to: options.to,
-        value: options.value,
-        callData: options.callData,
-      });
+      >(
+        '/userOp/signTransaction/hash',
+        {
+          accountId: options.accountId,
+          to: options.to,
+          value: options.value,
+          callData: options.callData,
+        },
+        config,
+      );
 
       return response.data.data;
     } catch (error) {