+++ src/account-abstraction/user-op/estimate/activate.ts
@@ -1,3 +1,5 @@
+import { AxiosRequestConfig } from 'axios';
+
 import { XellarEWBase } from '../../../base';
 import { BaseHttpResponse } from '../../../types/http';
 import { handleError, XellarError } from '../../../utils/error';
@@ -13,20 +15,27 @@ export class XellarAAEstimateActivate extends XellarEWBase {
    * @example
    *
    * ```typescript
-   * const result = await sdk.aa.estimate.activate({
+   * const result = await sdk.accountAbstraction.estimate.activate({
    *   accountId: "67b1f9310521667c3e94d625"
    * });
    * ```
    *
    * @see {@link https://docs.xellar.co/accountabstraction/api_reference/user_operation/estimate/activate/ Xellar Account Abstraction Estimate Activate Docs}
    */
-  async estimateActivate(options: EstimateActivateOptions) {
+  async estimateActivate(
+    options: EstimateActivateOptions,
+    config?: AxiosRequestConfig,
+  ) {
     try {
       const response = await this.aaInstance.post<
         BaseHttpResponse<EstimateResponse>
-      >('/userOp/deploy/estimate', {
-        accountId: options.accountId,
-      });
+      >(
+        '/userOp/deploy/estimate',
+        {
+          accountId: options.accountId,
+        },
+        config,
+      );
 
       return response.data.data;
     } catch (error) {