+++ src/account-abstraction/user-op/estimate/sign-transaction.ts
@@ -1,3 +1,5 @@
+import { AxiosRequestConfig } from 'axios';
+
 import { XellarEWBase } from '../../../base';
 import { BaseHttpResponse } from '../../../types/http';
 import { handleError, XellarError } from '../../../utils/error';
@@ -19,7 +21,7 @@ export class XellarAAEstimateSignTransaction extends XellarEWBase {
    * @example
    *
    * ```typescript
-   * const result = await sdk.aa.estimate.signTransaction({
+   * const result = await sdk.accountAbstraction.estimate.signTransaction({
    *   accountId: "67959f7927373a6808679de2",
    *   to: "0x...",
    *   value: "0",
@@ -29,16 +31,23 @@ export class XellarAAEstimateSignTransaction extends XellarEWBase {
    *
    * @see {@link https://docs.xellar.co/accountabstraction/api_reference/user_operation/estimate/sign_transaction/ Xellar Account Abstraction Estimate Sign Transaction Docs}
    */
-  async estimateSignTransaction(options: EstimateSignTransactionOptions) {
+  async estimateSignTransaction(
+    options: EstimateSignTransactionOptions,
+    config?: AxiosRequestConfig,
+  ) {
     try {
       const response = await this.aaInstance.post<
         BaseHttpResponse<EstimateResponse>
-      >('/userOp/signTransaction/estimate', {
-        accountId: options.accountId,
-        to: options.to,
-        value: options.value,
-        callData: options.callData,
-      });
+      >(
+        '/userOp/signTransaction/estimate',
+        {
+          accountId: options.accountId,
+          to: options.to,
+          value: options.value,
+          callData: options.callData,
+        },
+        config,
+      );
 
       return response.data.data;
     } catch (error) {