+++ src/account-abstraction/user-op/submit.ts
@@ -1,3 +1,5 @@
+import { AxiosRequestConfig } from 'axios';
+
 import { XellarEWBase } from '../../base';
 import { BaseHttpResponse } from '../../types/http';
 import { handleError, XellarError } from '../../utils/error';
@@ -16,7 +18,7 @@ export class XellarAASubmit extends XellarEWBase {
    * @example
    *
    * ```typescript
-   * const result = await sdk.aa.submitUserOp({
+   * const result = await sdk.accountAbstraction.submitUserOp({
    *   signature: "0x...",
    *   userOpId: "67b1fc9a442f26832dadd881",
    *   hash: "0x...",
@@ -26,16 +28,23 @@ export class XellarAASubmit extends XellarEWBase {
    *
    * @see {@link https://docs.xellar.co/accountabstraction/api_reference/user_operation/submit/ Xellar Account Abstraction Submit User Operation Docs}
    */
-  async submitUserOp(options: AASubmitUserOpOptions) {
+  async submitUserOp(
+    options: AASubmitUserOpOptions,
+    config?: AxiosRequestConfig,
+  ) {
     try {
       const response = await this.aaInstance.post<
         BaseHttpResponse<AASubmitUserOpResponse>
-      >('/userOp/submit', {
-        signature: options.signature,
-        userOpId: options.userOpId,
-        hash: options.hash,
-        isSponsored: options.isSponsored ?? false,
-      });
+      >(
+        '/userOp/submit',
+        {
+          signature: options.signature,
+          userOpId: options.userOpId,
+          hash: options.hash,
+          isSponsored: options.isSponsored ?? false,
+        },
+        config,
+      );
 
       return response.data.data;
     } catch (error) {